pyproject.toml
[tool.poetry]
name = "fast-api-chess-server"
version = "0.1.0"
description = ""
authors = ["Lucas Roig <lucas1.roig@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.62.0"
uvicorn = "^0.12.3"
chess = "^1.3.0"
loguru = "^0.5.3"
asyncpg = "^0.21.0"
databases = "^0.4.1"
alembic = "^1.4.3"
psycopg2 = "^2.8.6"
bcrypt = "^3.2.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^1.7.3"}
PyJWT = "^1.7.1"
pytest-asyncio = "^0.14.0"
asgi-lifespan = "^1.0.1"
httpx = "^0.16.1"
python-multipart = "^0.0.5"
python-jose = "^3.2.0"
six = "^1.16.0"
pytest-mock = "^3.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
